refactor(drive): 抽离上传队列行模型
